+++ rust/src/api/noise.rs
@@ -1,4 +1,4 @@
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 
 use dashmap::DashMap;
 pub use ipp::model::PrinterState;
@@ -6,17 +6,17 @@ pub use ipp::model::JobState;
 use turborand::rng::Rng;
 
 use crate::{hardware_control::live_view::white_noise::{self, WhiteNoiseGeneratorHandle}, helpers::log_debug, models::images::RawImage, utils::flutter_texture::FlutterTexture};
- 
+
 lazy_static::lazy_static! {
-    pub static ref NOISE_HANDLES: DashMap<usize, WhiteNoiseGeneratorHandle> = DashMap::<usize, WhiteNoiseGeneratorHandle>::new();
+    pub static ref NOISE_HANDLES: DashMap<u32, WhiteNoiseGeneratorHandle> = DashMap::<u32, WhiteNoiseGeneratorHandle>::new();
 }
 
-static NOISE_HANDLE_COUNT: AtomicUsize = AtomicUsize::new(1);
+static NOISE_HANDLE_COUNT: AtomicU32 = AtomicU32::new(1);
 
-const NOISE_WIDTH: usize = 1280;
-const NOISE_HEIGHT: usize = 720;
+const NOISE_WIDTH: u32 = 1280;
+const NOISE_HEIGHT: u32 = 720;
 
-pub fn noise_open(texture_ptr: usize) -> usize {
+pub fn noise_open(texture_ptr: usize) -> u32 {
     // Initialize noise and push noise frames to Flutter texture
     let renderer_main = FlutterTexture::new(texture_ptr, NOISE_WIDTH, NOISE_HEIGHT);
     let join_handle = white_noise::start_and_get_handle(NOISE_WIDTH, NOISE_HEIGHT, move |raw_frame| {
@@ -34,7 +34,7 @@ pub fn noise_get_frame() -> RawImage {
     white_noise::generate_frame(&Rng::new(), NOISE_WIDTH, NOISE_HEIGHT)
 }
 
-pub fn noise_close(handle_id: usize) {
+pub fn noise_close(handle_id: u32) {
     // Retrieve handle
     let handle = NOISE_HANDLES.remove(&handle_id).expect("Invalid noise handle ID");
