+++ rust/src/api/nokhwa.rs
@@ -4,22 +4,22 @@ pub fn nokhwa_get_cameras() -> Vec<NokhwaCameraInfo> {
     nokhwa::get_cameras()
 }
 
-pub fn nokhwa_open_camera(friendly_name: String, operations: Vec<ImageOperation>, texture_ptr: usize) -> usize {
+pub fn nokhwa_open_camera(friendly_name: String, operations: Vec<ImageOperation>, texture_ptr: usize) -> u32 {
     nokhwa::nokhwa_open_camera(friendly_name, operations, texture_ptr)
 }
 
-pub fn nokhwa_set_operations(handle_id: usize, operations: Vec<ImageOperation>) {
+pub fn nokhwa_set_operations(handle_id: u32, operations: Vec<ImageOperation>) {
     nokhwa::nokhwa_set_operations(handle_id, operations)
 }
 
-pub fn nokhwa_get_camera_status(handle_id: usize) -> CameraState {
+pub fn nokhwa_get_camera_status(handle_id: u32) -> CameraState {
     nokhwa::nokhwa_get_camera_status(handle_id)
 }
 
-pub fn nokhwa_get_last_frame(handle_id: usize) -> Option<RawImage> {
+pub fn nokhwa_get_last_frame(handle_id: u32) -> Option<RawImage> {
     nokhwa::nokhwa_get_last_frame(handle_id)
 }
 
-pub fn nokhwa_close_camera(handle_id: usize) {
+pub fn nokhwa_close_camera(handle_id: u32) {
     nokhwa::nokhwa_close_camera(handle_id)
 }
